backend/requirements-minimal.txt
# Core Flask dependencies - install these first
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Werkzeug==2.3.7
PyJWT==2.8.0
python-dotenv==1.0.0
bcrypt==4.1.2
SQLAlchemy==2.0.23